Added BackIssues font & the tone background
Added the Back Issues font locally and imported it, the folder location is in the app folder for ease of access. Applied the Back Issues font to the statement boxes as well.

Added a tone background to add a bit of texture to the website without it being distracting for the user.