Paper JS Integration
Recreated the CreateCavnasTools applicaiton in the PaperJS framework.  This allows for the easy ability to create layered canvases for the user to draw on.  This version has reached feature parity with the previous canvas component with the addition of switching layers as well as the ability to toggle their visibility.  The only downside to this is the bugged eraser that appears to draw to the canvas on any layer higher than the lowest layer in the hierarchy