3 Panel Editing
The user is now able to swap between 3 different panels to edit.  On switching panels the user's previous work is stored so that it may be recalled if they decide to work on that panel again.  When pressing save the stored data is saved into localStorage so that is can be recalled whenever they reenter the page